backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
pydantic
python-dotenv
slowapi
